requirements.txt
openai
streamlit
langchain
langchain_community
langchain-openai
pypdf
faiss-cpu
flask
langchain_google_community
chromadb
langchain_chroma
bs4
html2text
fpdf2
unidecode